Refactor AI services
