+++ src/services/ai-summary.test.ts
@@ -1,0 +1,3 @@
+describe('Test AI summary', () => {
+    test('');
+});
